Refactor: 예외 처리 추가

+++ src/main/java/com/gifthub/cart/controller/CartController.java
@@ -5,9 +5,12 @@ import com.gifthub.cart.service.CartService;
 import com.gifthub.gifticon.dto.GifticonDto;
 import com.gifthub.gifticon.service.GifticonService;
 import com.gifthub.global.exception.ExceptionResponse;
+import com.gifthub.point.exception.NotFoundGifticonException;
 import com.gifthub.user.UserJwtTokenProvider;
 import com.gifthub.user.dto.UserDto;
 import com.gifthub.user.exception.IdMismatchException;
+import com.gifthub.user.exception.NotFoundUserException;
+import com.gifthub.user.exception.NotLoginedException;
 import com.gifthub.user.service.UserService;
 import lombok.AllArgsConstructor;
 import org.apache.tools.ant.taskdefs.condition.Http;
@@ -32,12 +35,13 @@ public class CartController {
         Long userId = null;
         try {
             userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
-        } catch (Exception e) {
-            return ResponseEntity.badRequest().body("로그인을 해주세요.");
-        }
 
-        if (userId == null) {
-            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
+            if (isNull(userId)) {
+                throw new NotLoginedException();
+            }
+
+        } catch (NotLoginedException e) {
+            return ResponseEntity.badRequest().body(exceptionResponse.getException(e.getField(), e.getCode(), e.getMessage()));
         }
 
         return ResponseEntity.ok(cartService.findByUser(userId));
@@ -49,15 +53,23 @@ public class CartController {
             Long userId = null;
             UserDto userDto = null;
 
-            try {
-                userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
+            userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
 
-                userDto = userService.getUserById(userId);
-            } catch (Exception e) {
-                return ResponseEntity.badRequest().body("로그인을 해주세요.");
+            if (isNull(userId)) {
+                throw new NotLoginedException();
+            }
+
+            userDto = userService.getUserById(userId);
+
+            if (isNull(userDto)) {
+                throw new NotFoundUserException();
             }
 
             GifticonDto gifticonDto = gifticonService.findGifticon(gifticonId);
+
+            if (isNull(gifticonDto)) {
+                throw new NotFoundGifticonException();
+            }
 
             CartDto cartDto = CartDto.builder()
                     .userDto(userDto)
@@ -67,10 +79,8 @@ public class CartController {
             Long cartId = cartService.addToCart(cartDto);
 
             return ResponseEntity.ok().body(cartId);
-        } catch (Exception exception) {
-            exception.printStackTrace();
-
-            return ResponseEntity.badRequest().build();
+        } catch (NotLoginedException | NotFoundUserException | NotFoundGifticonException e) {
+            return ResponseEntity.badRequest().body(exceptionResponse.getException(e.getField(), e.getCode(), e.getMessage()));
         }
     }
 
@@ -100,14 +110,24 @@ public class CartController {
         try {
             Long userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
 
-            cartService.removeAllFromCart(userId);
-        } catch (Exception e) {
-            e.printStackTrace();
+            if (isNull(userId)) {
+                throw new NotLoginedException();
+            }
 
-            return ResponseEntity.badRequest().build();
+            cartService.removeAllFromCart(userId);
+        } catch (NotLoginedException e) {
+            return ResponseEntity.badRequest().body(exceptionResponse.getException(e.getField(), e.getCode(), e.getMessage()));
         }
 
         return ResponseEntity.ok().build();
+    }
+
+    private static <T> boolean isNull(T t) {
+        if (t == null) {
+            return true;
+        } else {
+            return false;
+        }
     }
 
 }
