+++ src/main/java/com/gifthub/gifticon/controller/GifticonController.java
@@ -5,8 +5,10 @@ import com.gifthub.gifticon.service.GifticonImageService;
 import com.gifthub.gifticon.service.GifticonService;
 import com.gifthub.gifticon.service.GifticonStorageService;
 import com.gifthub.gifticon.service.OcrService;
+import com.gifthub.global.exception.ExceptionResponse;
 import com.gifthub.product.service.ProductService;
 import com.gifthub.user.UserJwtTokenProvider;
+import com.gifthub.user.exception.NotLoginedException;
 import com.gifthub.user.service.UserService;
 import jakarta.servlet.ServletOutputStream;
 import jakarta.servlet.http.HttpServletResponse;
@@ -26,6 +28,7 @@ import java.io.IOException;
 public class GifticonController {
     private final GifticonService gifticonService;
     private final UserJwtTokenProvider userJwtTokenProvider;
+    private final ExceptionResponse exceptionResponse;
 
 
     @GetMapping("/barcode/{barcode}")
@@ -55,13 +58,15 @@ public class GifticonController {
         try {
             Long userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
 
+            if (isNull(userId)) {
+                throw new NotLoginedException();
+            }
+
             Page<GifticonDto> gifticons = gifticonService.getGifticonByUserId(pageable, userId);
 
             return ResponseEntity.ok(gifticons);
-        } catch (Exception e) {
-            e.printStackTrace();
-
-            return ResponseEntity.badRequest().build();
+        } catch (NotLoginedException e) {
+            return ResponseEntity.badRequest().body(exceptionResponse.getException(e.getField(), e.getCode(), e.getMessage()));
         }
 
     }
@@ -69,30 +74,37 @@ public class GifticonController {
     @PostMapping("/gifticon/delete/{gifticonId}")
     public ResponseEntity<Object> deleteGifticon(@PathVariable("gifticonId") Long gifticonId,
                                                  @RequestHeader HttpHeaders headers) {
-        Long userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
+        try {
+            Long userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
 
-        GifticonDto gifticonDto = gifticonService.findGifticon(gifticonId);
+            if (isNull(userId)) {
+                throw new NotLoginedException();
+            }
 
-        gifticonService.deleteById(gifticonId);
+            gifticonService.deleteById(gifticonId);
 
-        return ResponseEntity.ok().build();
+            return ResponseEntity.ok().build();
+        } catch (NotLoginedException e) {
+            return ResponseEntity.badRequest().body(exceptionResponse.getException(e.getField(), e.getCode(), e.getMessage()));
+        }
     }
 
     @PostMapping("/gifticon/forSale/{gifticonId}")
     public ResponseEntity<Object> setSale(@PathVariable("gifticonId") Long gifticonId,
                                           @RequestHeader HttpHeaders headers) {
-        Long userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
-
-        GifticonDto gifticonDto = gifticonService.findGifticon(gifticonId);
-
         try {
+            Long userId = userJwtTokenProvider.getUserIdFromToken(headers.get("Authorization").get(0));
+
+            if (isNull(userId)) {
+                throw new NotLoginedException();
+            }
+
             gifticonService.setSale(gifticonId);
 
             return ResponseEntity.ok().build();
-        } catch (Exception e) {
-            e.printStackTrace();
 
-            return ResponseEntity.badRequest().build();
+        } catch (NotLoginedException e) {
+            return ResponseEntity.badRequest().body(exceptionResponse.getException(e.getField(), e.getCode(), e.getMessage()));
         }
 
     }
@@ -100,14 +112,16 @@ public class GifticonController {
     @GetMapping("/gifticon/products/{productId}")
     public ResponseEntity<Object> findGifticonByProudctId(Pageable pageable,
                                                      @PathVariable("productId") Long productId) {
-        try {
-            Page<GifticonDto> gifticons = gifticonService.getGifticonByProudctId(pageable, productId);
+        Page<GifticonDto> gifticons = gifticonService.getGifticonByProudctId(pageable, productId);
 
-            return ResponseEntity.ok(gifticons);
-        } catch (Exception e) {
-            e.printStackTrace();
+        return ResponseEntity.ok(gifticons);
+    }
 
-            return ResponseEntity.badRequest().build();
+    private static <T> boolean isNull(T t) {
+        if (t == null) {
+            return true;
+        } else {
+            return false;
         }
     }
